+++ Code/Tools/AssetProcessor/native/utilities/UnitTestShaderCompilerServer.cpp
@@ -49,7 +49,7 @@ void UnitTestShaderCompilerServer::startServer()
 {
     if (!m_server->isListening())
     {
-        if (!m_server->listen(QHostAddress(m_serverAddress), m_serverPort))
+        if (!m_server->listen(QHostAddress(m_serverAddress), static_cast<quint16>(m_serverPort)))
         {
             AZ_TracePrintf(AssetProcessor::DebugChannel, "Server %s could not start.\n", m_serverAddress.toUtf8().data());
             emit errorMessage("Server could not start ");
