+++ Gems/Atom/RHI/Metal/Code/Source/RHI/SwapChain.h
@@ -53,7 +53,7 @@ namespace AZ
             id<MTLDevice> m_mtlDevice = nil;
             NativeWindowType* m_nativeWindow = nullptr;
             AZStd::vector<id<CAMetalDrawable>> m_drawables;
-            float m_refreshRate = 0.0f;
+            uint32_t m_refreshRate = 0;
         };
     }
 }
