+++ Gems/EMotionFX/Code/Tests/Vector2ToVector3CompatibilityTests.cpp
@@ -46,8 +46,8 @@ namespace EMotionFX
             m_blendTree->AddChildNode(finalNode);
             finalNode->AddUnitializedConnection(m_blendNNode, BlendTreeBlendNNode::PORTID_OUTPUT_POSE, BlendTreeFinalNode::PORTID_INPUT_POSE);
 
-            const int motionNodeCount = 3;
-            for (int i = 0; i < motionNodeCount; ++i)
+            const uint16 motionNodeCount = 3;
+            for (uint16 i = 0; i < motionNodeCount; ++i)
             {
                 AnimGraphMotionNode* motionNode = aznew AnimGraphMotionNode();
                 m_blendTree->AddChildNode(motionNode);
