+++ Gems/ScriptCanvasDeveloper/Code/Editor/Source/EditorAutomation/EditorAutomationActions/EditorKeyActions.cpp
@@ -30,7 +30,7 @@ namespace ScriptCanvasDeveloper
 #if defined(AZ_COMPILER_MSVC)
         INPUT osInput = { 0 };
         osInput.type = INPUT_KEYBOARD;
-        osInput.ki.wVk = m_keyValue;
+        osInput.ki.wVk = static_cast<WORD>(m_keyValue);
 
         switch (m_keyAction)
         {
